Include webhooks calls in more benchmark tests
