+++ saleor/graphql/checkout/tests/benchmark/test_homepage.py
@@ -1,42 +1,6 @@
-from unittest.mock import MagicMock
-
 import pytest
 
-from saleor.app.models import App
-from saleor.webhook.event_types import WebhookEventType
-from saleor.webhook.models import Webhook
-
 from ....tests.utils import get_graphql_content
-
-
-@pytest.fixture
-def mock_webhook_plugin_with_shipping_app(
-    settings,
-    permission_manage_checkouts,
-    monkeypatch,
-):
-    # Mock http requests as we are focusing on testing database access
-    response = MagicMock()
-    response.json.return_value = {"excluded_methods": []}
-    monkeypatch.setattr("requests.post", lambda *args, **kwargs: response)
-
-    # Enable webhook plugin
-    settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
-
-    # Create multiple apps with a shipping webhook
-    for i in range(3):
-        app = App.objects.create(name=f"Benchmark App {i}", is_active=True)
-        app.tokens.create(name="Default")
-        app.permissions.add(permission_manage_checkouts)
-        webhook = Webhook.objects.create(
-            name="shipping-webhook-1",
-            app=app,
-            target_url="https://shipping-gateway.com/api/",
-        )
-        webhook.events.create(
-            event_type=WebhookEventType.CHECKOUT_FILTER_SHIPPING_METHODS,
-            webhook=webhook,
-        )
 
 
 @pytest.mark.django_db
